dtype int32 the observation space

+++ gym_fuzz1ng/envs/fuzz_libpng_env.py
@@ -49,7 +49,7 @@ class FuzzLibPNGEnv(gym.Env):
             gym_fuzz1ng.libpng_target_path(), launch_afl_fuzz=True,
         )
         self.observation_space = gym.spaces.Box(
-            0, np.inf, shape=(2, coverage.PATH_MAP_SIZE), dtype='float32',
+            0, np.inf, shape=(2, coverage.PATH_MAP_SIZE), dtype='int32',
         )
         self.action_space = spaces.Discrete(_dict.size())
         self.reset()
